feat(viewer): move to Annotation style type type directory
