+++ packages/insight-viewer/src/Viewer/AnnotationDrawer/AnnotationDrawer.styles.ts
@@ -1,4 +1,4 @@
-import { AnnotationStyle } from '../types'
+import { AnnotationStyle } from '../../types'
 
 export const svgStyle: AnnotationStyle = {
   default: {
