+++ packages/insight-viewer/src/Viewer/types.ts
@@ -1,4 +1,3 @@
-import { CSSProperties } from 'react'
 import { CornerstoneImage } from '../utils/cornerstoneHelper'
 import { ProgressComponent, Viewport, OnViewportChange } from '../types'
 import { SetFrame } from '../hooks/useMultipleImages/useFrame'
@@ -18,9 +17,4 @@ export type ViewerProp = {
   onViewportChange?: OnViewportChange
   onFrameChange?: SetFrame
   interaction?: Interaction
-}
-
-export type AnnotationStyleType = 'default' | 'focus' | 'outline' | 'highlight'
-export type AnnotationStyle = {
-  [styleType in AnnotationStyleType]?: CSSProperties
 }
